Fixed lastCall errors associated with dplyr::last()
Removed dplyr::last() from the lastCall variable.
Subset the core_otus that pass the 'increase' threshold using a vector of OTU string identifiers. This is what the original function was doing, but dplyr:last() seemed to erroneously allow all OTUs to pass the filter regardless of the BC ranked value.